// packages/core/src/models/ConfigFiles.ts
/**
 * A ConfigFile is a known file that is used to customize
 * launch or build behaviour of the project.
 *
 * Examples of ConfigFiles are:
 * - boot.js
 * - webpack.config.js
 */
export type ConfigFileInfo = {

	/** Human readable ID of this file */
	slug: string;

	/** Path of the default file */
	defaultPath: string;

	/** Find and include from "bluerain" directory of project */
	findInBlueRain: boolean;

	/** Find and include files from bluerain plugins (apps and platforms) */
	findInPlugins: boolean;

	/** Is this a directory. Set true, otherwise false of files. */
	isDir: boolean;

	/** Does this file contain a hook function? */
	isHook: boolean,

	/** File name or pattern */
	name: string;

	/** The directory to look in */
	dir: string;

	/** Custom method to call when all files are resolved */
	handler?: (files: string[]) => void;
};